# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "NewsItemSummary": {
        "properties": {
          "canonical_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Canonical Url"
          },
          "editorial_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Editorial Score"
          },
          "fetched_at": {
            "title": "Fetched At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "market": {
            "title": "Market",
            "type": "string"
          },
          "published_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "sentiment_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sentiment Label"
          },
          "source_name": {
            "title": "Source Name",
            "type": "string"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "source_name",
          "market",
          "fetched_at"
        ],
        "title": "NewsItemSummary",
        "type": "object"
      },
...
      "QuoteSummaryView": {
        "properties": {
          "change_amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Change Amount"
          },
          "change_percent": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Change Percent"
          },
          "day_high": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Day High"
          },
          "day_low": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Day Low"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "fetched_at": {
            "title": "Fetched At",
            "type": "string"
          },
          "has_hot_alert": {
            "default": false,
            "title": "Has Hot Alert",
            "type": "boolean"
          },
          "market": {
            "title": "Market",
            "type": "string"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "open_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Open Price"
          },
          "previous_close": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previous Close"
          },
          "price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price"
          },
          "provider_symbol": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Symbol"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "status": {
            "default": "unavailable",
            "enum": [
              "ok",
              "delayed",
              "unavailable",
              "symbol_not_supported",
              "fetch_failed"
            ],
            "title": "Status",
            "type": "string"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          },
          "volume": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Volume"
          }
        },
        "required": [
          "symbol",
          "market",
          "fetched_at"
        ],
        "title": "QuoteSummaryView",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "WatchlistAiInsightView": {
        "properties": {
          "failover": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failover"
          },
          "generated_at": {
            "title": "Generated At",
            "type": "string"
          },
          "insight_text": {
            "title": "Insight Text",
            "type": "string"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          }
        },
        "required": [
          "symbol",
          "insight_text",
          "generated_at"
        ],
        "title": "WatchlistAiInsightView",
        "type": "object"
      },
      "WatchlistCandidateView": {
        "properties": {
          "aliases": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Aliases",
            "type": "array"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "market": {
            "title": "Market",
            "type": "string"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          }
        },
        "required": [
          "symbol",
          "market",
          "display_name"
        ],
        "title": "WatchlistCandidateView",
        "type": "object"
      },
      "WatchlistItemCreate": {
        "properties": {
          "alert_mode": {
            "default": "fixed",
            "title": "Alert Mode",
            "type": "string"
          },
          "alert_threshold": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alert Threshold"
          },
          "display_name": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Display Name",
            "type": "string"
          },
          "market": {
            "title": "Market",
            "type": "string"
          },
          "symbol": {
            "maxLength": 32,
            "minLength": 1,
            "title": "Symbol",
            "type": "string"
          }
        },
        "required": [
          "symbol",
          "market",
          "display_name"
        ],
        "title": "WatchlistItemCreate",
        "type": "object"
      },
      "WatchlistItemView": {
        "properties": {
          "alert_mode": {
            "title": "Alert Mode",
            "type": "string"
          },
          "alert_threshold": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alert Threshold"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "market": {
            "title": "Market",
            "type": "string"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          }
        },
        "required": [
          "id",
          "symbol",
          "market",
          "display_name",
          "is_active",
          "alert_mode"
        ],
        "title": "WatchlistItemView",
        "type": "object"
      },
      "WatchlistResearchBriefView": {
        "properties": {
          "drivers": {
            "items": {
              "$ref": "#/components/schemas/WatchlistResearchDriverView"
            },
            "title": "Drivers",
            "type": "array"
          },
          "generated_at": {
            "title": "Generated At",
            "type": "string"
          },
          "has_unexplained_price_move": {
            "default": false,
            "title": "Has Unexplained Price Move",
            "type": "boolean"
          },
          "market": {
            "title": "Market",
            "type": "string"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          },
          "top_action_level": {
            "enum": [
              "act_now",
              "watch_today",
              "know_only",
              "none"
            ],
            "title": "Top Action Level",
            "type": "string"
          },
          "window_days": {
            "default": 14,
            "title": "Window Days",
            "type": "integer"
          }
        },
        "required": [
          "symbol",
          "market",
          "generated_at",
          "top_action_level"
        ],
        "title": "WatchlistResearchBriefView",
        "type": "object"
      },
      "WatchlistResearchDriverView": {
        "properties": {
          "action_level": {
            "enum": [
              "act_now",
              "watch_today",
              "know_only"
            ],
            "title": "Action Level",
            "type": "string"
          },
          "category": {
            "enum": [
              "policy_macro",
              "company_action",
              "supply_chain",
              "price_action"
            ],
            "title": "Category",
            "type": "string"
          },
          "news_item": {
            "$ref": "#/components/schemas/NewsItemSummary"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "category",
          "action_level",
          "reason",
          "news_item"
        ],
        "title": "WatchlistResearchDriverView",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "News Caught Backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/market/watchlist": {
      "get": {
        "operationId": "list_watchlist_quotes_api_market_watchlist_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-App-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-App-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/QuoteSummaryView"
                  },
                  "title": "Response List Watchlist Quotes Api Market Watchlist Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Watchlist Quotes",
        "tags": [
          "market"
        ]
      }
    },
...
    "/api/watchlist": {
      "get": {
        "operationId": "list_watchlist_api_watchlist_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-App-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-App-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WatchlistItemView"
                  },
                  "title": "Response List Watchlist Api Watchlist Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Watchlist",
        "tags": [
          "watchlist"
        ]
      },
      "post": {
        "operationId": "create_watchlist_item_api_watchlist_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-App-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-App-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WatchlistItemCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchlistItemView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Watchlist Item",
        "tags": [
          "watchlist"
        ]
      }
    },
    "/api/watchlist/candidates": {
      "get": {
        "operationId": "get_watchlist_candidates_api_watchlist_candidates_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-App-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-App-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WatchlistCandidateView"
                  },
                  "title": "Response Get Watchlist Candidates Api Watchlist Candidates Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Watchlist Candidates",
        "tags": [
          "watchlist"
        ]
      }
    },
    "/api/watchlist/{symbol}": {
      "delete": {
        "operationId": "delete_watchlist_item_api_watchlist__symbol__delete",
        "parameters": [
          {
            "in": "path",
            "name": "symbol",
            "required": true,
            "schema": {
              "title": "Symbol",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-App-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-App-Token"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Watchlist Item",
        "tags": [
          "watchlist"
        ]
      }
    },
    "/api/watchlist/{symbol}/ai-insight": {
      "post": {
        "operationId": "get_watchlist_ai_insight_api_watchlist__symbol__ai_insight_post",
        "parameters": [
          {
            "in": "path",
            "name": "symbol",
            "required": true,
            "schema": {
              "title": "Symbol",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-App-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-App-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchlistAiInsightView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Watchlist Ai Insight",
        "tags": [
          "watchlist"
        ]
      }
    },
    "/api/watchlist/{symbol}/related-news": {
      "get": {
        "operationId": "list_related_news_api_watchlist__symbol__related_news_get",
        "parameters": [
          {
            "in": "path",
            "name": "symbol",
            "required": true,
            "schema": {
              "title": "Symbol",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-App-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-App-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/NewsItemSummary"
                  },
                  "title": "Response List Related News Api Watchlist  Symbol  Related News Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Related News",
        "tags": [
          "watchlist"
        ]
      }
    },
    "/api/watchlist/{symbol}/research-brief": {
      "get": {
        "operationId": "get_watchlist_research_brief_api_watchlist__symbol__research_brief_get",
        "parameters": [
          {
            "in": "path",
            "name": "symbol",
            "required": true,
            "schema": {
              "title": "Symbol",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-App-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-App-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchlistResearchBriefView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Watchlist Research Brief",
        "tags": [
          "watchlist"
        ]
      }
    },
...
  }
}